Code cleanup and refactoring started
